links provided for settings view items
